+++ Changelog.md
@@ -9,6 +9,7 @@
 - Added: OpenHashSets now implement foreach and have less overhead.
 - Added: ImmutableOpenHashSet that is not editable (is linked by default for fast iteration)
 - Added: CustomOpenHashSets now implement foreach and have less overhead.
+- Added: ImmutableOpenHashMap that is not editable (is linked by default for fast iteration)
 
 ### Version 0.3.1
 - Fixed: containsKey & containsValue in HashMaps were deprecated for Object Variants.